Environ list type
The environschema is an abstraction (unwanted/un-needed?) for creating
fields for a restricted set. With that in mind, we also need to start
exposing a list type so we gain a similar feature to controller
features.

The code is rather simple as we expose and existing type `schema.List`
and the tests just ensure it works as expected.